+++ FitWorkout/backend/controllers/WorkoutController.php
@@ -2,17 +2,15 @@
 
 namespace backend\controllers;
 
+use common\models\Exercise;
 use common\models\Workout;
+use common\models\Workoutexercise;
+use Yii;
 use yii\data\ActiveDataProvider;
+use yii\filters\AccessControl;
+use yii\filters\VerbFilter;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
-use yii\filters\VerbFilter;
-use common\models\Exercise;
-use common\models\Exercisecategory;
-use common\models\Exercisetype;
-use common\models\Workoutexercise;
-use yii\helpers\ArrayHelper;
-use yii\filters\AccessControl;
 
 /**
  * WorkoutController implements the CRUD actions for Workout model.
@@ -59,7 +57,7 @@ class WorkoutController extends Controller
      */
     public function actionIndex()
     {
-        if (\Yii::$app->user->can('admin')) {
+        if (Yii::$app->user->can('admin')) {
 
             $dataProvider = new ActiveDataProvider([
                 'query' => Workout::find(),
@@ -78,9 +76,9 @@ class WorkoutController extends Controller
             return $this->render('index', [
                 'dataProvider' => $dataProvider,
             ]);
-        } else if (\Yii::$app->user->can('personalTrainer')) {
+        } else if (Yii::$app->user->can('personalTrainer')) {
             $dataProvider = new ActiveDataProvider([
-                'query' => Workout::find()->where(['ptId' => \Yii::$app->user->id]),
+                'query' => Workout::find()->where(['ptId' => Yii::$app->user->id]),
                 /*
             'pagination' => [
                 'pageSize' => 50
@@ -97,6 +95,8 @@ class WorkoutController extends Controller
                 'dataProvider' => $dataProvider,
             ]);
         }
+
+        return 0;
     }
 
     /**
@@ -107,19 +107,37 @@ class WorkoutController extends Controller
      */
     public function actionView($id)
     {
-        $exerciseWorkoutList =  Workoutexercise::find()->where(['workoutId' => $id])->all();
-        $exerciseList = [];
-        foreach ($exerciseWorkoutList as $exerciseWorkout) {
-            $exerciseList[] = $exerciseWorkout->exerciseId;
+        /** Verificar permissão do utilizador */
+        if (Yii::$app->user->can('readWorkoutsAdmin')) {
+            $exerciseWorkoutList = Workoutexercise::find()->where(['workoutId' => $id])->all();
+            $exerciseList = [];
+            foreach ($exerciseWorkoutList as $exerciseWorkout) {
+                $exerciseList[] = $exerciseWorkout->exerciseId;
+            }
+            $exerciseModel = Exercise::find()->where(['id' => $exerciseList])->all();
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+                'exerciseModel' => $exerciseModel,
+            ]);
         }
 
-        $exerciseModel = Exercise::find()->where(['id' => $exerciseList])->all();
+        return 0;
+    }
 
+    /**
+     * Finds the Workout model based on its primary key value.
+     * If the model is not found, a 404 HTTP exception will be thrown.
+     * @param int $id ID
+     * @return Workout the loaded model
+     * @throws NotFoundHttpException if the model cannot be found
+     */
+    protected function findModel($id)
+    {
+        if (($model = Workout::findOne($id)) !== null) {
+            return $model;
+        }
 
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-            'exerciseModel' => $exerciseModel,
-        ]);
+        throw new NotFoundHttpException('The requested page does not exist.');
     }
 
     /**
@@ -129,35 +147,41 @@ class WorkoutController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Workout();
+        /** Verificar permissão do utilizador */
+        if (Yii::$app->user->can('createWorkoutAdmin')) {
+            $model = new Workout();
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post())) {
+                    $model->ptId = Yii::$app->user->id;
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post())) {
-                $model->ptId = \Yii::$app->user->id;
-
-                $model->save();
-                return $this->redirect(['view', 'id' => $model->id]);
+                    $model->save();
+                    return $this->redirect(['view', 'id' => $model->id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
+            return $this->render('create', [
+                'model' => $model,
+            ]);
         }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+        return 0;
     }
 
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        /** Verificar permissão do utilizador */
+        if (Yii::$app->user->can('updateWorkoutsAdmin')) {
+            $model = $this->findModel($id);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        return 0;
     }
 
     /**
@@ -169,9 +193,13 @@ class WorkoutController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        /** Verificar permissão do utilizador */
+        if (Yii::$app->user->can('deleteWorkoutsAdmin')) {
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        }
 
-        return $this->redirect(['index']);
+        return 0;
     }
 
     public function actionExercisesadd($id)
@@ -196,8 +224,8 @@ class WorkoutController extends Controller
     {
         $model = new Workoutexercise();
 
-        $model->exerciseId = \Yii::$app->request->post('exerciseId', null);;
-        $model->workoutId = \Yii::$app->request->post('workoutId', null);;
+        $model->exerciseId = Yii::$app->request->post('exerciseId', null);
+        $model->workoutId = Yii::$app->request->post('workoutId', null);
         $model->exerciseCalories = 10;
         $model->equipmentWeight = 10;
         $model->seriesSize = 3;
@@ -207,21 +235,5 @@ class WorkoutController extends Controller
         if ($model->save()) {
             return $this->redirect(['/workout']);
         }
-    }
-
-    /**
-     * Finds the Workout model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param int $id ID
-     * @return Workout the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
-    {
-        if (($model = Workout::findOne($id)) !== null) {
-            return $model;
-        }
-
-        throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
